+++ CHANGES.md
@@ -15,6 +15,7 @@ Version 0.6.0 (Pending)
 - Only show badges user has permission to see during at-con check in
 - Added support for setting manual price at con with the attendee_override_price right
 - Added support for Till Name saved by cookie client-side and saved when till session is closed (schema change)
+- Set till name to "Attendee Import" during attendee import
 
 Version 0.5 (4/9/2019)
 - Initial "release" version. Not ready for production yet, but start tracking changes